requirements.txt
python-binance
krakenex
pykrakenapi
pandas
